requirements.txt
aniso8601==9.0.1
blinker==1.4
click==7.1.2
colorama==0.4.4
Deprecated==1.2.13
Flask==1.1.4
flask-buzz==0.1.15
Flask-Limiter==1.4
Flask-Mail==0.9.1
flask-praetorian==1.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
limits==1.5.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
passlib==1.7.4
pendulum==2.1.2
py-buzz==1.0.3
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytzdata==2020.1
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==1.0.1
wrapt==1.13.2
